+++ packages/blockchain-wallet-v4-frontend/src/scenes/SecurityCenter/Advanced/TwoStepVerificationRemember/Settings/template.js
