+++ packages/blockchain-wallet-v4-frontend/src/scenes/SecurityCenter/Advanced/TwoStepVerificationRemember/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 
 import { selectors } from 'data'
-import TwoStepVerification from './template.js'
+import TwoStepVerification from './template'
 
 class TwoStepVerificationRememberContainer extends React.PureComponent {
   render () {
@@ -11,7 +11,7 @@ class TwoStepVerificationRememberContainer extends React.PureComponent {
 }
 
 const mapStateToProps = (state) => ({
-  authTypeNeverSave: selectors.core.settings.getAuthTypeNeverSave(state)
+  authTypeNeverSave: selectors.core.settings.getAuthTypeNeverSave(state).getOrElse(0)
 })
 
 export default connect(mapStateToProps)(TwoStepVerificationRememberContainer)
